integration/pyproject.toml
[project]
name = "integration"
version = "0.1.0"
description = "Integration test for thermocktat"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pymodbus>=3.11.4",
    "paho-mqtt>=2.1.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.14",
    "ty>=0.0.14",
]
